refactor: Maven 多模块改造 + 共享库 + 测试文件整理
目录结构重构为 Maven 多模块项目：
- java-services/pom.xml（父 POM，统一版本管理）
- shared/common-models（共享 DTO，HealthResponse）
- shared/poi-utils（共享 POI 工具，DocxHelper）
- poi-service（引用 shared 模块）

变更：
- DocxHelper 从 poi-service/util/ 移至 shared/poi-utils/
- poi-service pom.xml 改为继承父 POM，版本由父 POM 管理
- 测试文件从 poi-service/ 移至 java-services/tests/
- CLAUDE.md 移至根目录
- 新增 java-services/.gitignore（排除 target/ 和测试输出）

验证：mvnw clean compile 4 模块全部通过，13/13 E2E 测试通过
