+++ java-services/.gitignore
@@ -1,0 +1,17 @@
+# Build artifacts
+target/
+*.class
+*.jar
+!**/maven-wrapper.jar
+
+# Test output
+test_output/
+test_output_*.docx
+test_output_archive/
+
+# IDE
+.idea/
+*.iml
+.settings/
+.classpath
+.project
