+++ java-services/poi-service/pom.xml
@@ -5,23 +5,14 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.5.14</version>
-        <relativePath/>
+        <groupId>com.fachuan</groupId>
+        <artifactId>java-services</artifactId>
+        <version>1.0.0</version>
     </parent>
 
-    <groupId>com.fachuan</groupId>
     <artifactId>poi-service</artifactId>
-    <version>1.0.0</version>
     <name>POI Document Service</name>
     <description>Apache POI document generation service for Fachuan</description>
-
-    <properties>
-        <java.version>25</java.version>
-        <poi.version>5.5.1</poi.version>
-        <poi-ooxml.version>5.5.1</poi-ooxml.version>
-    </properties>
 
     <dependencies>
         <dependency>
@@ -29,32 +20,34 @@
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
 
-        <!-- Apache POI core -->
+        <!-- Shared modules -->
+        <dependency>
+            <groupId>com.fachuan</groupId>
+            <artifactId>common-models</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.fachuan</groupId>
+            <artifactId>poi-utils</artifactId>
+        </dependency>
+
+        <!-- Apache POI (versions managed by parent) -->
         <dependency>
             <groupId>org.apache.poi</groupId>
             <artifactId>poi</artifactId>
-            <version>${poi.version}</version>
         </dependency>
-
-        <!-- Apache POI OOXML (for .docx, .xlsx, .pptx) -->
         <dependency>
             <groupId>org.apache.poi</groupId>
             <artifactId>poi-ooxml</artifactId>
-            <version>${poi-ooxml.version}</version>
         </dependency>
-
-        <!-- Apache POI OOXML full schemas (for CTP/CTSectPr OOXML types) -->
         <dependency>
             <groupId>org.apache.poi</groupId>
             <artifactId>poi-ooxml-full</artifactId>
-            <version>${poi-ooxml.version}</version>
         </dependency>
 
-        <!-- Lombok for reducing boilerplate -->
+        <!-- Lombok -->
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
-            <optional>true</optional>
         </dependency>
 
         <!-- Jackson for JSON -->
@@ -73,19 +66,6 @@
 
     <build>
         <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <annotationProcessorPaths>
-                        <path>
-                            <groupId>org.projectlombok</groupId>
-                            <artifactId>lombok</artifactId>
-                            <version>1.18.38</version>
-                        </path>
-                    </annotationProcessorPaths>
-                </configuration>
-            </plugin>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
