+++ java-services/shared/poi-utils/src/main/java/com/fachuan/shared/poi/DocxHelper.java
@@ -1,4 +1,4 @@
-package com.fachuan.poi.util;
+package com.fachuan.shared.poi;
 
 import org.apache.poi.util.Units;
 import org.apache.poi.wp.usermodel.HeaderFooterType;
