+++ java-services/tests/test_poi_integration.py
